# restaurant-management-api-contract.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Restaurant Management System - Unified API",
    "version": "1.0.0",
    "description": "Consolidated API contract for all microservices via API Gateway"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "API Gateway"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders": {
      "get": {
        "tags": [
          "order-controller"
        ],
        "summary": "List all orders in the system",
        "operationId": "getAllOrders",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "order-controller"
        ],
        "summary": "Place a new guest order",
        "operationId": "placeOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Order"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "tags": [
          "order-controller"
        ],
        "summary": "View order details by ID",
        "operationId": "getOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "order-controller"
        ],
        "summary": "Cancel a pending order",
        "operationId": "cancelOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/health": {
      "get": {
        "tags": [
          "order-controller"
        ],
        "operationId": "health",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/restaurants/tables": {
      "get": {
        "tags": [
          "Restaurant Management"
        ],
        "summary": "View all tables",
        "operationId": "getTables",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RestaurantTable"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Restaurant Management"
        ],
        "summary": "Add a new physical table to the restaurant",
        "operationId": "addTable",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestaurantTable"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RestaurantTable"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/restaurants/reservations": {
      "get": {
        "tags": [
          "Restaurant Management"
        ],
        "summary": "List all active reservations",
        "operationId": "getReservations",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Reservation"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Restaurant Management"
        ],
        "summary": "Create a new table reservation",
        "operationId": "createReservation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Reservation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Reservation"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/restaurants/menu": {
      "get": {
        "tags": [
          "Restaurant Management"
        ],
        "summary": "Get full menu",
        "operationId": "getMenu",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MenuItem"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Restaurant Management"
        ],
        "summary": "Add new menu item",
        "operationId": "addMenuItem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MenuItem"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/restaurants/inventory": {
      "get": {
        "tags": [
          "Restaurant Management"
        ],
        "summary": "List all inventory levels",
        "operationId": "getInventory",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InventoryItem"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Restaurant Management"
        ],
        "summary": "Restock inventory item",
        "operationId": "updateInventory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryItem"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryItem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/restaurants/tables/{id}/status": {
      "patch": {
        "tags": [
          "Restaurant Management"
        ],
        "summary": "Update table status (AVAILABLE/OCCUPIED)",
        "operationId": "updateTableStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RestaurantTable"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/restaurants/health": {
      "get": {
        "tags": [
          "Restaurant Management"
        ],
        "operationId": "health",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/kitchen/tickets/{id}/status": {
      "patch": {
        "tags": [
          "Kitchen Operations"
        ],
        "summary": "Update ticket preparation status",
        "operationId": "updateStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/KitchenOrder"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/kitchen/tickets": {
      "get": {
        "tags": [
          "Kitchen Operations"
        ],
        "summary": "Get all active kitchen tickets",
        "operationId": "getActiveTickets",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KitchenOrder"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/kitchen/tickets/{id}": {
      "get": {
        "tags": [
          "Kitchen Operations"
        ],
        "summary": "Get a kitchen ticket by ID",
        "operationId": "getTicket",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/KitchenOrder"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/delivery/tasks/{id}/dispatch": {
      "patch": {
        "tags": [
          "Delivery Management"
        ],
        "summary": "Mark task as DISPATCHED and assign courier",
        "operationId": "dispatchTask",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "courierName",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DeliveryTask"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/delivery/tasks/{id}/complete": {
      "patch": {
        "tags": [
          "Delivery Management"
        ],
        "summary": "Mark task as DELIVERED",
        "operationId": "completeTask",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DeliveryTask"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/delivery/tasks": {
      "get": {
        "tags": [
          "Delivery Management"
        ],
        "summary": "Get all delivery tasks",
        "operationId": "getAllTasks",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeliveryTask"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/delivery/tasks/{id}": {
      "get": {
        "tags": [
          "Delivery Management"
        ],
        "summary": "Get a delivery task by ID",
        "operationId": "getDeliveryTask",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DeliveryTask"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounting/invoices/{id}/pay": {
      "patch": {
        "tags": [
          "Accounting & Finance"
        ],
        "summary": "Process invoice payment",
        "operationId": "payInvoice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounting/invoices": {
      "get": {
        "tags": [
          "Accounting & Finance"
        ],
        "summary": "Get all invoices",
        "operationId": "getAllInvoices",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Invoice"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounting/invoices/{id}": {
      "get": {
        "tags": [
          "Accounting & Finance"
        ],
        "summary": "Get a bill (invoice) by ID",
        "operationId": "getBill",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "customerName": {
            "type": "string"
          },
          "restaurantId": {
            "type": "integer",
            "format": "int64"
          },
          "totalAmount": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "idempotencyKey": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          }
        }
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "menuItemId": {
            "type": "integer",
            "format": "int64"
          },
          "itemName": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number"
          }
        }
      },
      "RestaurantTable": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "restaurantId": {
            "type": "integer",
            "format": "int64"
          },
          "tableNumber": {
            "type": "integer",
            "format": "int32"
          },
          "seatingCapacity": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "Reservation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "tableId": {
            "type": "integer",
            "format": "int64"
          },
          "guestName": {
            "type": "string"
          },
          "contactNumber": {
            "type": "string"
          },
          "guestCount": {
            "type": "integer",
            "format": "int32"
          },
          "reservationTime": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          },
          "idempotencyKey": {
            "type": "string"
          }
        }
      },
      "MenuItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "InventoryItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "restaurantId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "quantity": {
            "type": "number",
            "format": "double"
          },
          "unit": {
            "type": "string"
          },
          "minThreshold": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "KitchenOrder": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "orderId": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DeliveryTask": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "orderId": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string"
          },
          "courierName": {
            "type": "string"
          },
          "deliveryAddress": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Invoice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "orderId": {
            "type": "integer",
            "format": "int64"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "paidAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}